+++ TelCo.ColorCoder/Main.cs
@@ -13,30 +13,30 @@ class MainClass
  private static void Main(string[] args)
         {
             int pairNumber = 4;
-            ColorPair testPair1 = GetNumberAndColorPair.GetColorFromPairNumber(pairNumber);
+            ColorPair testPair1 = GetColorPair.GetColorFromPairNumber(pairNumber);
             Console.WriteLine("[In]Pair Number: {0},[Out] Colors: {1}\n", pairNumber, testPair1);
             Debug.Assert(testPair1.majorColor == Color.White);
             Debug.Assert(testPair1.minorColor == Color.Brown);
 
             pairNumber = 5;
-            testPair1 = GetNumberAndColorPair.GetColorFromPairNumber(pairNumber);
+            testPair1 = GetColorPair.GetColorFromPairNumber(pairNumber);
             Console.WriteLine("[In]Pair Number: {0},[Out] Colors: {1}\n", pairNumber, testPair1);
             Debug.Assert(testPair1.majorColor == Color.White);
             Debug.Assert(testPair1.minorColor == Color.SlateGray);
 
             pairNumber = 23;
-            testPair1 = GetNumberAndColorPair.GetColorFromPairNumber(pairNumber);
+            testPair1 = GetColorPair.GetColorFromPairNumber(pairNumber);
             Console.WriteLine("[In]Pair Number: {0},[Out] Colors: {1}\n", pairNumber, testPair1);
             Debug.Assert(testPair1.majorColor == Color.Violet);
             Debug.Assert(testPair1.minorColor == Color.Green);
 
             ColorPair testPair2 = new ColorPair() { majorColor = Color.Yellow, minorColor = Color.Green };
-            pairNumber = GetNumberAndColorPair.GetPairNumberFromColor(testPair2);
+            pairNumber = GetNumber.GetPairNumberFromColor(testPair2);
             Console.WriteLine("[In]Colors: {0}, [Out] PairNumber: {1}\n", testPair2, pairNumber);
             Debug.Assert(pairNumber == 18);
 
             testPair2 = new ColorPair() { majorColor = Color.Red, minorColor = Color.Blue };
-            pairNumber = GetNumberAndColorPair.GetPairNumberFromColor(testPair2);
+            pairNumber = GetNumber.GetPairNumberFromColor(testPair2);
             Console.WriteLine("[In]Colors: {0}, [Out] PairNumber: {1}", testPair2, pairNumber);
             Debug.Assert(pairNumber == 6);
         }
